cleanup(core): refactor nx list command to remove unused code and code duplication
