+++ packages/nx/src/command-line/list/list.ts
@@ -1,21 +1,17 @@
-import { workspaceRoot } from '../../utils/workspace-root';
-import { output } from '../../utils/output';
-import {
-  fetchCorePlugins,
-  getInstalledPluginsAndCapabilities,
-  listCorePlugins,
-  listInstalledPlugins,
-  listPluginCapabilities,
-} from '../../utils/plugins';
-import {
-  getLocalWorkspacePlugins,
-  listLocalWorkspacePlugins,
-} from '../../utils/plugins/local-plugins';
+import { readNxJson } from '../../config/nx-json';
 import {
   createProjectGraphAsync,
   readProjectsConfigurationFromProjectGraph,
 } from '../../project-graph/project-graph';
-import { readNxJson } from '../../config/nx-json';
+import { output } from '../../utils/output';
+import {
+  getInstalledPluginsAndCapabilities,
+  getLocalWorkspacePlugins,
+  listAlsoAvailableCorePlugins,
+  listPluginCapabilities,
+  listPlugins,
+} from '../../utils/plugins';
+import { workspaceRoot } from '../../utils/workspace-root';
 
 export interface ListArgs {
   /** The name of an installed plugin to query  */
@@ -31,14 +27,13 @@ export interface ListArgs {
  *
  */
 export async function listHandler(args: ListArgs): Promise<void> {
-  const nxJson = readNxJson();
   const projectGraph = await createProjectGraphAsync({ exitOnError: true });
   const projects = readProjectsConfigurationFromProjectGraph(projectGraph);
 
   if (args.plugin) {
     await listPluginCapabilities(args.plugin, projects.projects);
   } else {
-    const corePlugins = fetchCorePlugins();
+    const nxJson = readNxJson();
 
     const localPlugins = await getLocalWorkspacePlugins(projects, nxJson);
     const installedPlugins = await getInstalledPluginsAndCapabilities(
@@ -47,22 +42,20 @@ export async function listHandler(args: ListArgs): Promise<void> {
     );
 
     if (localPlugins.size) {
-      listLocalWorkspacePlugins(localPlugins);
+      listPlugins(localPlugins, 'Local workspace plugins:');
     }
-    listInstalledPlugins(installedPlugins);
-    listCorePlugins(installedPlugins, corePlugins);
+    listPlugins(installedPlugins, 'Installed plugins:');
+    listAlsoAvailableCorePlugins(installedPlugins);
 
     output.note({
       title: 'Community Plugins',
       bodyLines: [
         'Looking for a technology / framework not listed above?',
         'There are many excellent plugins maintained by the Nx community.',
-        'Search for the one you need here: https://nx.dev/plugins/registry.',
+        'Search for the one you need here: https://nx.dev/plugin-registry.',
       ],
     });
 
-    output.note({
-      title: `Use "nx list [plugin]" to find out more`,
-    });
+    output.note({ title: `Use "nx list [plugin]" to find out more` });
   }
 }
