+++ packages/nx/src/utils/plugins/installed-plugins.ts
@@ -1,15 +1,14 @@
-import * as chalk from 'chalk';
-import { output } from '../output';
-import type { PluginCapabilities } from './models';
-import { getPluginCapabilities } from './plugin-capabilities';
-import { hasElements } from './shared';
-import { readJsonFile } from '../fileutils';
-import { PackageJson, readModulePackageJson } from '../package-json';
-import { workspaceRoot } from '../workspace-root';
 import { join } from 'path';
 import { readNxJson } from '../../config/nx-json';
-import { getNxRequirePaths } from '../installation-directory';
 import { ProjectConfiguration } from '../../config/workspace-json-project-json';
+import { readJsonFile } from '../fileutils';
+import { getNxRequirePaths } from '../installation-directory';
+import { PackageJson, readModulePackageJson } from '../package-json';
+import { workspaceRoot } from '../workspace-root';
+import {
+  PluginCapabilities,
+  getPluginCapabilities,
+} from './plugin-capabilities';
 
 export function findInstalledPlugins(): PackageJson[] {
   const packageJsonDeps = getDependenciesFromPackageJson();
@@ -91,32 +90,4 @@ export async function getInstalledPluginsAndCapabilities(
   }
 
   return result;
-}
-
-export function listInstalledPlugins(
-  installedPlugins: Map<string, PluginCapabilities>
-) {
-  const bodyLines: string[] = [];
-
-  for (const [, p] of installedPlugins) {
-    const capabilities = [];
-    if (hasElements(p.executors)) {
-      capabilities.push('executors');
-    }
-    if (hasElements(p.generators)) {
-      capabilities.push('generators');
-    }
-    if (p.projectGraphExtension) {
-      capabilities.push('graph-extensions');
-    }
-    if (p.projectInference) {
-      capabilities.push('project-inference');
-    }
-    bodyLines.push(`${chalk.bold(p.name)} (${capabilities.join()})`);
-  }
-
-  output.log({
-    title: `Installed plugins:`,
-    bodyLines: bodyLines,
-  });
 }
