+++ packages/nx/src/utils/plugins/plugin-capabilities.ts
@@ -1,19 +1,22 @@
-import * as chalk from 'chalk';
 import { dirname, join } from 'path';
-
+import {
+  ExecutorsJsonEntry,
+  GeneratorsJsonEntry,
+} from '../../config/misc-interfaces';
 import { ProjectConfiguration } from '../../config/workspace-json-project-json';
-import { NxPlugin, readPluginPackageJson } from '../../project-graph/plugins';
-import { loadNxPlugin } from '../../project-graph/plugins/loader';
 import { readJsonFile } from '../fileutils';
 import { getNxRequirePaths } from '../installation-directory';
-import { output } from '../output';
+import { NxPlugin, readPluginPackageJson } from '../../project-graph/plugins';
+import { loadNxPlugin } from '../../project-graph/plugins/loader';
 import { PackageJson } from '../package-json';
-import { getPackageManagerCommand } from '../package-manager';
-import { workspaceRoot } from '../workspace-root';
-import { hasElements } from './shared';
 
-import type { PluginCapabilities } from './models';
-import type { ExecutorsJsonEntry } from '../../config/misc-interfaces';
+export interface PluginCapabilities {
+  name: string;
+  executors?: { [name: string]: ExecutorsJsonEntry };
+  generators?: { [name: string]: GeneratorsJsonEntry };
+  projectInference?: boolean;
+  projectGraphExtension?: boolean;
+}
 
 function tryGetCollection<T extends object>(
   packageJsonPath: string,
@@ -120,146 +123,4 @@ async function tryGetModule(
   } catch {
     return null;
   }
-}
-
-export async function listPluginCapabilities(
-  pluginName: string,
-  projects: Record<string, ProjectConfiguration>
-) {
-  const plugin = await getPluginCapabilities(
-    workspaceRoot,
-    pluginName,
-    projects
-  );
-
-  if (!plugin) {
-    const pmc = getPackageManagerCommand();
-    output.note({
-      title: `${pluginName} is not currently installed`,
-      bodyLines: [
-        `Use "${pmc.addDev} ${pluginName}" to install the plugin.`,
-        `After that, use "${pmc.exec} nx g ${pluginName}:init" to add the required peer deps and initialize the plugin.`,
-      ],
-    });
-
-    return;
-  }
-
-  const hasBuilders = hasElements(plugin.executors);
-  const hasGenerators = hasElements(plugin.generators);
-  const hasProjectGraphExtension = !!plugin.projectGraphExtension;
-  const hasProjectInference = !!plugin.projectInference;
-
-  if (
-    !hasBuilders &&
-    !hasGenerators &&
-    !hasProjectGraphExtension &&
-    !hasProjectInference
-  ) {
-    output.warn({ title: `No capabilities found in ${pluginName}` });
-    return;
-  }
-
-  const bodyLines = [];
-
-  if (hasGenerators) {
-    bodyLines.push(chalk.bold(chalk.green('GENERATORS')));
-    bodyLines.push('');
-    bodyLines.push(
-      ...Object.keys(plugin.generators).map(
-        (name) => `${chalk.bold(name)} : ${plugin.generators[name].description}`
-      )
-    );
-    if (hasBuilders) {
-      bodyLines.push('');
-    }
-  }
-
-  if (hasBuilders) {
-    bodyLines.push(chalk.bold(chalk.green('EXECUTORS/BUILDERS')));
-    bodyLines.push('');
-    bodyLines.push(
-      ...Object.keys(plugin.executors).map(
-        (name) =>
-          `${chalk.bold(name)} : ${resolveExecutorDescription(
-            pluginName,
-            plugin.executors[name],
-            projects
-          )}`
-      )
-    );
-  }
-
-  if (hasProjectGraphExtension) {
-    bodyLines.push(`✔️  Project Graph Extension`);
-  }
-
-  if (hasProjectInference) {
-    bodyLines.push(`✔️  Project Inference`);
-  }
-
-  output.log({
-    title: `Capabilities in ${plugin.name}:`,
-    bodyLines,
-  });
-}
-
-function resolveExecutorDescription(
-  pluginName: string,
-  executorJsonEntry: ExecutorsJsonEntry,
-  projects: Record<string, ProjectConfiguration>,
-  requirePaths = getNxRequirePaths(workspaceRoot)
-) {
-  try {
-    if (typeof executorJsonEntry === 'string') {
-      // it points to another executor, resolve it
-      const [pkgName, executor] = executorJsonEntry.split(':');
-      // read the package.json of the parent plugin
-      const { path: packageJsonPath } = readPluginPackageJson(
-        pluginName,
-        projects,
-        requirePaths
-      );
-      // accumulate the require paths to resolve nested packages
-      const cummulativeRequirePaths = [
-        ...requirePaths,
-        dirname(packageJsonPath),
-      ];
-      const collection = loadExecutorsCollection(
-        pkgName,
-        projects,
-        cummulativeRequirePaths
-      );
-
-      return resolveExecutorDescription(
-        pkgName,
-        collection[executor],
-        projects,
-        cummulativeRequirePaths
-      );
-    }
-
-    return executorJsonEntry.description;
-  } catch {
-    return 'No description available';
-  }
-}
-
-function loadExecutorsCollection(
-  pluginName: string,
-  projects: Record<string, ProjectConfiguration>,
-  requirePaths: string[]
-): { [name: string]: ExecutorsJsonEntry } {
-  const { json: packageJson, path: packageJsonPath } = readPluginPackageJson(
-    pluginName,
-    projects,
-    requirePaths
-  );
-
-  return {
-    ...tryGetCollection(packageJsonPath, packageJson.builders, 'builders'),
-    ...tryGetCollection(packageJsonPath, packageJson.executors, 'builders'),
-    ...tryGetCollection(packageJsonPath, packageJson.builders, 'executors'),
-    ...tryGetCollection(packageJsonPath, packageJson.executors, 'executors'),
-  };
 }
